新增权限 auth_system 标识系统内置变量，被标识变量禁止增加删除子项操作

+++ string.h
@@ -9,8 +9,8 @@
 #ifndef __nosym__
 
 u32 vname_check(char *name);
-u32 vname_gen(unsigned char *name);
-u32 vhead_gen(u64 head);
+u64 vname_gen(unsigned char *name);
+u64 vhead_gen(u64 head);
 u64 vname_head(unsigned char *name);
 char* skip_string(char *exp);
 char* get_string(char *exp, char **pexp);
